MB-13827 Update weight ticket updater used for ghc weight ticket test
Need to use the office one instead of the customer one.
Also update the weight ticket and ppm shipment created to match the expected state when the office goes in to edit.
These came up as I was testing if partial updates could be made in any of the ghc ppm doc update endpoints and while I'm not fixing the partial updates issue in this PR, I figured I'd keep these changes since I have the tweaks already.