MB-13827 Fix PPM make funcs to update PPMShipment pointer on Shipment
This is only needed for some PPM make functions, the ones that call another func to make the shipment and then make edits to the returned shipment. The reason is that we're getting a value back from the make funcs instead of a pointer, so every change we make after that to the PPM shipment is no longer reflected in the pointer to the PPM shipment that the MTO shipment has (i.e. `ppmShipment.Shipment.PPMShipment` no longer points at the latest copy of the PPMShipment). A more thorough fix would be to update all the PPM make funcs to return pointers, but since they'll eventually get replaced anyway with [our new testdatagen pattern](https://transcom.github.io/mymove-docs/docs/backend/testing/factory-test-data-generation/), just doing a quick fix here.