Adding logging to the ClientCertFromContext;
This adds logging to CloudWatch which will help to understand if there's
an issue when the reading for a client certificate rather than a session
in the Middleware package for RequestLogger. There may be some massaging
that I will need to do further here to test things a bit better.