Adding a migration for adding Daycos and MoveHQ;
This patch adds the necessary steps to add certificates for 3rd party
users of the Prime API. The main work is done in two parts. There are
also two different client certificates we are adding in this migration.
The first one is for DAYCOS and the other is for MOVEHQ.

This migration adds a user without a login.gov ID to the `users` table.
This user is created by has never logged into the system. This userID
is then used as a foreign key in the `client_certs` table.

Co-Authored-By: Chikara T <chikara@truss.works>