Updating the Admin API metadata;
This touches the schemes, host, and contact stanzas only. The main
reason for using the email address for Pagerduty is this is what the
client is expecting to use for support escalation. We can modify this
email to be a placeholder `admin@example.com` email as well. I've also
made the `host:` stanza equal to the production URL as Swagger 2.x
doesn't allow for multiple hosts the way OpenAPI 3.x does. Since the
main place this API will be used is in the PRD environment, I don't
think this is a problem.