Sorting the defintions;
We can most likely turn this into an error again if we want. This only
removes a single warning from the checks, but I think it's important to
have these sorted in our `swagger-def` directory as it's now much easier
to read all the definitions and know what we support when code-folding
is enabled in your editor of choice.