Updating the handler and services;
This update is necessary since the Swagger and generated Go code was
updated. This was mostly done with my editor's LSP and not manually. I
was able to target the errors across files and run a find-and-replace to
ensure there were no errors. Once that was done, I ran the `make
server_test` to ensure that tests passed.