Adding descriptions & markdown refactor;
This patch is moving the summary stanza over to the description stanza.
This doesn't fix all of the issues though as descriptions need to have
at least 128 characters in them to be compliant. This patch also adds
references to Markdown files to make updating these kinds of
descriptions much easier for all Trussels on the project.