refactor (api): fix typos & move two summaries to be w/descriptions
The other summary stanzas had already been moved to be next to the description stanzas. The two in this commit are just ones I noticed slipped through when reviewing the PR.
One of the descriptions was for a `get` request, but described a `post` request, so I updated the wording to match properly.

JIRA ticket: MB-14723