feat(prime): :construction: Store origin duty location GBLOC on the orders table
For consistency and historical data accuracy, move the origin duty location gbloc to a static value on the orders record rather than a calculated value from a view. Drop the view and backfill the new column. Ensure that new orders created have correct GBLOC assigned, and that testdatagen orders assign correct value as well. Includes logic for USMC.

Migration script introduces changes to the schema: Add a column and Drop a view