refactor (cli): make db ping failure be fatal for ppm upload converter
I'd copied the logic from the server where apparently it's ok to not be able to connect yet (I'm guessing because other set up is happening) and apparently the failure messes things up when deploying containers. For the ppm command though, it seems better to just end the process instead.

JIRA Ticket: MB-15340