fix (payload): fix conversion of ppm shipment payload for creation
Previously we were setting secondary zips to a blank string on creation which then got saved to the DB. For updates, we use the same func but our service object takes care of converting that blank string pointer to `nil`. An alternative way of fixing this instead of having two separate conversion functions would be to add something to the ppm creator service object to change blank strings to nil. Not sure what is better tbh.

JIRA Ticket: MB-15488