Moving `config/` directory & modifying pre-commit;
This is needed because without it, CRA eject complains about there
already being a `config` file or directory. I also needed to modify the
pre-commit configuration because it explicitly calls out the `config/`
directory. No other modifications were made here, so I'm expecting to
revert this so I don't have to make the rest of the codebase aware that
the `config/` directory is now `config.milmove/`.