Fix Upload Search
There were two problems:
- we got a JS error when searching for a bad ID, rather than displaying our own error message
- when searching for an existing ID, we would get a bunch of console errors and nothing
  would be displayed

I added keys to the list of `TextField` elements to get rid of some of our console errors.
I also had to change things around to pass render props down from ShowController. I don't
totally understand why this was necessary but it helped.

To fix the empty page, I had to switch <ShowView> with <Show>. Not super sure why
but I noticed <ShowView> wasn't been used anywhere else in the app, and other
uses of SimpleShowLayout were all wrapped in Show.
This allowed the fields to render on the page, but about half of them were
empty because we aren't really getting the upload record from context.
So I changed to get that from the controllerProps, and that allowed
all the fields to be populated.