feat (service): add helper func to convert a file to PDF
This is a private function that will be called by a new service object. This part can be a standalone func since it will be needed repeatedly in that service object and is a connection point with gotenberg. This also makes it easier to test in isolation from the rest of the service object.

JIRA Ticket: MB-15488