Use first contract code instead of hardcoded one
This is a proof of concept.
Rather than using the hard-coded contract code, this changes pricing code
to grab a contract code from the database. This essentially picks a random
contract code, so it's not the final version of the business logic, but
it surfaces places that need to change.

In particular, this highlights the fact that we don't usually make a contract
record when creating data in our tests, so lots of tests had to be changed.
There's probably a slicker way to put this into testdatagen/factories so we
don't need to modify every one of these tests.