feat (services): Add service member and orders to PPMShipmentFetcher
This gets us some of the data we'll need for the AOA and Payment packets. Tbh, not sure if this should go here, or if the packet service objects should be using a move/order fetcher in addition to the ppm shipment fetcher. Only reason I chose this for now is that we don't have generic fetchers for those two as far as I know, and it's out of scope to do that too.

refactor (services tests): update the `PPMShipmentFetcher` tests that use `suite.False` and `suite.True` to give more helpful error messages

JIRA Ticket: MB-15488