fix (tests): fix flaky tests that were comparing pdf byte slices
We have the tests to try to ensure we're getting back the PDFs we expect, but since they are PDFs, we can't really check the text in them that easily, so we're comparing the byte slices. @ahobson noted that it looked like a race condition because it was the expected bytes that were empty in the flaky tests, and suggested we move when we read the expected bytes.

JIRA Ticket: MB-15488
Co-authored-by: Andrew Hobson <ahobson@truss.works>