Add temporary server logs for GetOrdersHandler
I'm trying to figure out if we're passing something weird back from
GetOrdersHandler or if stuff is getting mangled by Swagger.

If some of these new logs pop up and we get a null 200 response, it indicates that
the problem is in our code.
If none of them pop up and we get a null 200 response, it indicates that the
issue might be happening in go-swagger code.

Note that the addition of responder.Payload.Validate could change the bug
behavior because it can panic. So it's possible that we will get an error instead
of a 200.