Reversing the SIT duration list on the backend;
Originally, I had been attempting to reverse the SIT duration / SIT
history list on the client-side but because of how coupled together
different SITDisplay components are on the client-side. So now I've
reversed the output of SITDurationUpdates at the handler level to ensure
that they will always be displayed correctly.