Updating the Delve debug command;
The previous iteration of the command didn't work for me and was written
a long time ago. This version works and allows for my DAP-client to
connect to it remotely.

DAP is an acronym for Debug Adapter Protocol. This is built into VSCode,
Goland, and can be accessed via other editors such as Vim with plugins.