Refactoring the return of SIT Status;
Now that SIT service items with a future date are returned in SIT
Status, we have to update how we calculate the SIT end date and take
into account the SIT start date. There's a chance here to consolidate
how we do this in the future, but for now it's copied and pasted across
all three client Components that use SIT end date.