Fix move.AvailableToPrimeAt in devseed data
when a move is approved, we put the current time in this field.
This timestamp is used to figure out the current contract.
Other timestamps on the move are based on testdatagen.GHCTestYear (2018)
The contract active in 2018 does't have a contract year for the current date,
so pricing would fail.

All the different times used when creating test moves feels pretty haphazard,
it would be great to have a more comprehensive plan for how we use time in tests