Use strings rather than booleans because POSIX;
Since these environment variables are running within a POSIX shell
instead of in Yaml themselves, the correct values supported here are
either strings or numbers. While this has been working, it's better to
be explicit here about these values. `SERVE_API_INTERNAL=true` is
equivalent to `SERVE_API_INTERNAL='true'` in either case, but with
quotes we ensure that we're creating variables in a POSIX compatible
way.