Smaller changes due to tooling;
A blank parameters object isn't supported in the Yaml schema and since
we don't use it, I've commented it out rather than deleting it. In case
we use it in the future, we have it there to remind folks that the
`paramter:` key should be added there.

The double-quotes were replaced here automatically by my Yaml formatter.

And last, there was an extra line before `integration_tests_load` that
was automatically formatted and removed.