chore: update react-router and react-router-dom as much as we can
We can't update to the latest versions because it breaks our react-admin site. Per their troubleshooting docs, https://marmelab.com/react-admin/Routing.html#troubleshooting, we need to match the versions they're using, so we need to set resolutions. A bit of a pain because now we'll need to keep them in sync.

JIRA Ticket: MB-15974