fix: update sample value in typescript docstring
For some reason it thought we were trying to use a number with a leading zero. It was a postal code with a leading zero, but the type is `string`, so I'm not sure why it's messing up. The value itself doesn't matter much though since it's just an example, so I just changed it to a postal code that doesn't have a leading zero.

JIRA Ticket: MB-15974