fix: send psql stdout/err to `/dev/null` when checking the db connection
We previously did this and it got changed unintentionally. For some reason, not sending this to ``/dev/null` can break `Makefile` run configurations that run targets that use the `wait-for-db` script. I didn't see anything that stood out as a way to fix that, so for now we'll go with the easier/faster route of reverting this change. If we do want to change how much info we get back in this script, we can dedicate more time to figuring out why the run configurations won't work.