Refactoring the Prime docs;
This will make modifying the Prime description text easier over all. I
still need to follow this patter for the other sections as well. But
this is a start. Also this fails to pass `markdown-lint` because of
internal links within the description file. I will need to consider how
to fix this in the follow-up commits.