fix: return error if user tries to upload amended orders for another sm
Previously we didn't check if the service member making the request matched the service member on the orders. We now check if the request is a customer app request (since we're doing this at a service object level), and if so, include their service member ID as part of the query. We'll return a 404 if it doesn't match, which is better security since we won't confirm for other users that they even have a valid order ID.

JIRA Ticket: MB-16179