Reject updates to delivery address for shipments with SIT
Found a service that checks for SIT on a shipment and had to update
the way the test data was constructed to let it find it.
I was constructing a DOASIT service item initially, but the
service doesn't check for that one.
I think that's fine because you'll never have DOASIT without
other SIT service items also being present.