Adding definitions for sitAddressUpdates;
In this patch, I'm adding some definitions for SIT Address Update and
SIT Address Updates. With this, Swagger will generate the necessary code
for me to be able to use these properties in the _Models to Payload_
in order to pass the SITAddressUpdaates from the database.

The thing is that I have not been able to successfully get the Go code
to compile even with these changes. I'll commit the next code without
verification to get it in the branch, but I'll need some help with this
tomorrow.