Adding in new UUIDs to avoid collisions;
This was first addressed in #11033, but while reviewing #11187 I ran into a UUID
collision again. I've posted in [#g-testing][slack]🔒. I'm looking for some
feedback here to see if this fix works for all the other `devseed` scenarios.
I'll be doing further testing on Monday and will document on this PR what I
find. I didn't include this as part of #11187 so that I don't hold up that work.

[slack]: https://ustcdp3.slack.com/archives/CNTE04R5Z/p1691793822591899