feat (service): update `GetPPMShipment` to account for customer authz
Should work the way it was before if the request is coming from the prime API or the office app. Only adds authz check if it is from the customer app. If it is a customer request, we ensure the service member in the session matches the one that the shipment belongs to (via orders association).

Updated `moving_expense_creator` and `progear_weight_ticket_creator` to call `GetPPMShipment` differently now that it has the logic to use the logged in customer info as appropriate.
Also updated `moving_expense_creator` to pass the errors from `GetPPMShipment` along rather than turning them into server errors.

JIRA-Ticket: MB-16387