docs (comment): remove comment about ways of setting up request bodies
Got alignment from several people on the team that using the `go-swagger` types for integration tests with routing & middleware enabled seems to be the best path forward in terms of dev ergonomics and type safety. Someone brought up a great point that if you change the type in the yaml file, this will also lead to a compilation error in the integration tests, which is a faster feedback loop than having to wait for the tests to run to see that the body is no longer being set up properly.

JIRA-Ticket: MB-16387