removed comments
The fixme comment was found to be functional. The commented out TODO test was for another approach of marking a past sit as current, but do to harmful readability and tech debt effects this was abandoned. It was only meant to fix how current sit info displays
on the dashboard, and this was acheived by showing the most recent pastSIT information if there is no current sit