B-22244 add playwright test to approve a Delivery Address Change on an iHHG,
specifically testing that the change is applied after saving the Review Modal,
no need to save on the Shipment Edit page too.

Fixed a bug in testharness/make_move.go, the MS service item must be on the Move not the Shipment.
Also fixed a typo in the dHHG test.