Initial INPK rate engine swappage. I need more coffee, golang has gotta be capable of this
Committing it because it is a nightmare to debug and this lets me submit it. Need to fix market factor lookup next and take a better look at how the rate engine is set up so I can make it a more clean approach instead of throwing swaps. This is what we have structs and interfaces for right? There has to be a cleaner approach. But this is also the safest approach for enforcing IHPK logic on INPK. I will find a way, I gotta find a way. Also why is the frontend saying its missing billable weight like what it is right there, it would fail to price if it did not have weight present. Gotta look into that