Merge remote-tracking branch 'origin/B-23603-main' into b-23713-upload-av-status-main
Did some additonal work in this merge with moving code from ghcapi to internalapi. This isn't just a typical merge this includes upstream work in the GHCAPI, where this ticket is moving said antivirus handlers from GHCAPI to internalapi. You'll see that work in here. I got the server tests fixed up, client tests are acting a little weird so I'm merging this in and then merging main in to see if that changes anything before finishing up the client changes. Also some client changes yes because the old upstreams didn't have a helper func to make using the SSE easier, so I merged some of that work too with some changes.